break off SuiteFormatsTest
