+++ corehq/apps/app_manager/tests/test_suite.py
@@ -271,34 +271,6 @@ class SuiteTest(SimpleTestCase, TestXmlMixin):
     def test_no_case_assertions(self):
         self._test_generic_suite('app_no_case_sharing', 'suite-no-case-sharing')
 
-    def _test_format(self, detail_format, template_form):
-        app = Application.wrap(self.get_json('app_audio_format'))
-        details = app.get_module(0).case_details
-        details.short.get_column(0).format = detail_format
-        details.long.get_column(0).format = detail_format
-
-        expected = """
-        <partial>
-          <template form="{0}">
-            <text>
-              <xpath function="picproperty"/>
-            </text>
-          </template>
-          <template form="{0}">
-            <text>
-              <xpath function="picproperty"/>
-            </text>
-          </template>
-        </partial>
-        """.format(template_form)
-        self.assertXmlPartialEqual(expected, app.create_suite(), "./detail/field/template")
-
-    def test_audio_format(self):
-        self._test_format('audio', 'audio')
-
-    def test_image_format(self):
-        self._test_format('picture', 'image')
-
     def test_attached_picture(self):
         self._test_generic_suite_partial('app_attached_image', "./detail", 'suite-attached-image')
